% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndex.R
\name{getBowtieIndex}
\alias{getBowtieIndex}
\title{Get path to bowtie index on Rosalind}
\usage{
getBowtieIndex(
  species = c("human", "mouse", "rat", "fly", "cyno5", "cyno6"),
  what = c("dna", "rna"),
  allele = c("ref", "maj", "min")
)
}
\arguments{
\item{species}{String specifying which species.
Must be either human (default), mouse, rat, fly or cyno.}

\item{what}{String specifying what type of index should
be returned. Either "dna" (default) or "rna".
For "dna", this corresponds to the reference genome.
For "rna", this corresponds to a transcriptome.}

\item{allele}{String specifying which allele should be
used for human DNA. Either "ref" (default),
"major" or "minor".}
}
\value{
String specifying path to a bowtie index
}
\description{
Get path to bowtie index on Rosalind
}
\author{
Jean-Philippe Fortin
}
