% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCasRxRfBinaries.R
\name{getCasRxRfBinaries}
\alias{getCasRxRfBinaries}
\title{Get paths of binaries for getCasRxRfScores on Rosalind}
\usage{
getCasRxRfBinaries()
}
\value{
Named character vector of length 3.
}
\description{
Get paths of binaries for getCasRxRfScores on Rosalind
}
\author{
Jean-Philippe Fortin
}
