% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChromatinFiles.R
\name{getChromatinFiles}
\alias{getChromatinFiles}
\title{Get chromatin files for human on Rosalind}
\usage{
getChromatinFiles(species = c("human", "mouse"))
}
\arguments{
\item{species}{String specifying which species.
Only human can be provided at the moment.}
}
\value{
Named character vector of length 3.
}
\description{
Get chromatin files for human on Rosalind
}
\author{
Jean-Philippe Fortin
}
