% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConservationFiles.R
\name{getConservationFiles}
\alias{getConservationFiles}
\title{Get conservation files for human on Rosalind}
\usage{
getConservationFiles(species = c("human", "mouse"))
}
\arguments{
\item{species}{String specifying which species.}
}
\value{
String specifying path of the BigWig file.
}
\description{
Get conservation files for human on Rosalind
}
\author{
Jean-Philippe Fortin
}
