% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenome.R
\name{getGenomeFasta}
\alias{getGenomeFasta}
\title{Get path to fasta file on Rosalind}
\usage{
getGenomeFasta(species = c("human", "mouse"), allele = c("ref", "maj", "min"))
}
\arguments{
\item{species}{String specifying which species.
Must be either human (default), mouse, rat or fly.}

\item{allele}{String specifying which allele should be
used for human DNA. Either "ref" (default),
"major" or "minor".}
}
\value{
String specifying path to a fasta file
}
\description{
Get path to fasta file on Rosalind
}
\author{
Jean-Philippe Fortin
}
