% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenome.R
\name{getGenomePackage}
\alias{getGenomePackage}
\title{Shortcut to get a default BSGenome object}
\usage{
getGenomePackage(species = c("human", "mouse"))
}
\arguments{
\item{species}{String specifying which species.
Must be either human (default), mouse, rat, fly or cyno.}
}
\value{
BSgenome object.
}
\description{
Shortcut to get a default BSGenome object
}
\author{
Jean-Philippe Fortin
}
